LPS-70620 Rename internal package
